(feature): add clustercatalog controller for managing default clustercatalogs
that ensures any ClusterCatalog resources in the catalogd operands are applied
to the cluster, allowing modifications to any fields that the manifests don't
express an opinion on.

ClusterCatalogs are managed by this controller by using server-side apply.
Any errors encountered during this process will result in marking the olm/cluster
object with a Degraded status for the ClusterCatalogController that encountered
the error.

Signed-off-by: everettraven <everettraven@gmail.com>