(bugfix): prevent discovery errors on unknown GVKs by
checking for known GVK to meta.RESTMapping conversions before using a live
client for discovery. This is particularly useful in the scenario we currently
have with catalogd including CRs before the CRDs have been applied to the
cluster meaning that discovery has no knowledge of the GVK or how to map it
to a particular GVR and scope.

Signed-off-by: everettraven <everettraven@gmail.com>