Add support for experimental manifests
Add support for experimental manifests in operator-controller and catalogd

When a new operator-controller or catalogd bundle has the experimental
manifests, there will be a `/cp-manifests` script in the container to
copy the images to appripriate (new) directories. If the `/cp-manifests`
script does not exist, then the original method of copying the manifests
is used.

When deciding upon which manifests to use, the code defaults to the
original location. If the `standard` manifest location is found, it is
used instead. If the `experimental` manifest location is found *and* the
FeatureGate/FeatureSet is TechPreview, then the experimental manifests
are used.

In addition, check for existing `--feature-gates` options and remove
them before adding TechPreview enabled gates. This may mean there is
a discrepency between the existing and new feature-gate flags. A
check will be done in an upcoming commit to report discrepencies.

Signed-off-by: Todd Short <todd.short@me.com>